\name{kerdiest-internal}
\alias{A1_k}
\alias{A2_k}
\alias{derivative_kernel_function}
\alias{derivative_normal_kernel}
\alias{kernel_distribution_without_i}
\alias{mu2_k}
\alias{dichotomy_fun}
\alias{functional}
\alias{ro_k}
\alias{kernel_function}
\alias{kernel_function_distribution}
\alias{simp_int} 
\title{Internal Grid Functions}
\description{
  Internal Grid functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }

