\name{kerdiest-package}
\alias{kerdiest-package}
\alias{kerdiest}
\docType{package}
\title{Nonparametric Kernel Estimation of the Distribution Function
}
\description{Nonparametric kernel distribution function estimation for
  continuous random variables is performed. Three automatic bandwidth
  selection procedures for nonparametric kernel distribution function
  estimation are implemented: the plug-in method of Altman and Leger, the
  plug-in method of Polansky and Baker, and the modified cross-validation
  method of Bowman, Hall and Prvan. The exceedance function, the mean
  return period and the return level are also computed.
}
\references{
  Quintela-del-Río, A. and Estévez-Pérez, G. (2012), "Nonparametric kernel
  distribution function estimation with kerdiest: an R package for bandwidth
  choice and applications", Journal of Statistical Software, 50(8), 1-21.
}
\author{Graciela Estévez Pérez and Alejandro Quintela del Río

Maintainer: Ignacio López-de-Ullibarri
}

