\name{kerdiest-package}
\alias{kerdiest-package}
\alias{kerdiest}
\docType{package}
\title{
Nonparametric kernel estimation of the distribution. Bandwidth selection methods 
and estimation of related functions.
}
\description{
Nonparametric kernel distribution function estimation for continuous random 
variables is performed. Three automatic bandwidth selection procedures for
nonparametric kernel distribution function estimation are implemented: the
plug-in method of Altman and Leger, the plug-in method of Polansky and Baker, 
and the modified cross-validation method of Bowman, Hall and Prvan. The 
exceedance function, the mean return period and the return level 
are also computed.
}

\author{
Graciela Estevez Perez \email{graci@udc.es} and
Alejandro Quintela del Rio \email{aquintela@udc.es}

Maintainer: Alejandro Quintela del Rio \email{aquintela@udc.es}
}

